+++ web/resources/assets/javascript/dashboard.js
@@ -9,11 +9,6 @@ var _ = require('./common/helper');
 var PieChart = ReactD3.PieChart;
 var LineChart = ReactD3.LineChart;
 
-var name = $('.environment-left').first().text();
-var value = $('.environment-right').first().text();
-var placeholdername = $('#inputenvname').val();
-var placeholdervalue = $('#inputenvvalue').val();
-
 class Charts extends React.Component {
 
   constructor(props) {
@@ -193,23 +188,6 @@ if (charts) {
     });
   });
 
-  var renderEnvVariable = function() {
-
-    var data = $('input[name="environment"]').val();
-
-    if (data.length) {
-      $('#table-environment').html('');
-      data = data.split(',');
-      data.forEach(function(item) {
-        var key = item.split('=')[0];
-        var val = item.split('=')[1];
-        addNewRow(key, val);
-      });
-
-    }
-
-  };
-
   $('#project-modal').on('show.bs.modal', function(e) {
     var relatedTarget = e.relatedTarget;
 
@@ -222,9 +200,6 @@ if (charts) {
     $('#project-modal').find('select').each(function(e, node) {
       node.selectedIndex = 1;
     });
-
-    $('#table-environment').html('');
-    addNewRow('', '');
 
     if (type !== 'edit') {
       dialogIsEidtMode = false;
@@ -255,8 +230,6 @@ if (charts) {
           $('textarea[name="description"]').val(data.description);
           $('input[name="repositoryUrl"]').val(data.repositoryUrl);
           $('input[name="repositoryBranch"]').val(data.repositoryBranch);
-          $('input[name="environment"]').val(data.environment);
-          renderEnvVariable();
           if (duration) {
             $('select[name="hours"]').val(hours);
             $('select[name="minutes"]').val(minutes);
@@ -271,27 +244,6 @@ if (charts) {
   });
 
   $('#submit-project').on('click', function() {
-
-    var envvalue;
-    var rows = $('#table-environment tr');
-    var tempArr = rows.map(function (idx, el) {
-      var inputs = $(el).find('input');
-      var k = inputs[0].value.trim();
-      var v = inputs[1].value.trim();
-      if (k === '' || k == null) {
-        return;
-      }
-      return k + '=' + v;
-    }).toArray().filter(function (item) {
-      if (item.indexOf('undefined') === -1) {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    envvalue = tempArr.join(',');
-    $('#envvalue').val(envvalue);
-
     var data = {};
     $('#project-modal').find('input').each(function(e, node) {
       data[node.name] = node.value;
@@ -307,8 +259,8 @@ if (charts) {
       data[node.name] = val;
     });
 
-    var verified = ['title', 'repositoryUrl', 'repositoryBranch', 'runiOS'].every(v => {
-      if (data[v] == null) {
+    var verified = ['title', 'repositoryUrl', 'repositoryBranch'].every(v => {
+      if (!data[v]) {
         $(`input[name=${v}]`).parent().addClass('alert-danger alert-dismissible');
         return false;
       } else {
@@ -330,7 +282,6 @@ if (charts) {
     $.ajax({
       url: '/api/project/' + (dialogIsEidtMode ? 'update' : 'add') + '?_csrf=' + csrf_token,
       data: data,
-      async: false,
       method: 'post',
       success: function(d) {
         if (d.success) {
@@ -381,20 +332,6 @@ if (charts) {
         location.reload();
       }
     });
-  });
-
-  var addNewRow = function (inputname, inputvalue) {
-    var newRow = '<tr> ' +
-      '<td class="td-environment"><span class="environment-left">' + name + '</span></td> ' +
-      '<td name="environment" class="td-environment"><input type="text" class="input-env" placeholder="' + placeholdername + '" value="' + inputname + '"/></td> ' +
-      '<td class="td-environment"><span class="environment-right">' + value + '</span></td> ' +
-      '<td name="environment" class="td-environment"><input type="text" class="input-env" placeholder="' + placeholdervalue + '" value="' + inputvalue + '"/></td> ' +
-      '</tr>';
-    $('#table-environment').append(newRow);
-  };
-
-  $('#add-environment').on('click', function () {
-    addNewRow('', '');
   });
 
   $('.dropdown-menu a').on('click', function(e) {